minor cleanup in dft code

+++ src/dft_func/XCIntegratorForGGA.cpp
@@ -837,7 +837,7 @@ integrateFxcFockForGgaClosedShell(const std::vector<double*>&       aoFockPointe
 
     timer.stop("Preparation");
 
-    for (int box_id = 0; box_id < counts.size(); box_id++)
+    for (size_t box_id = 0; box_id < counts.size(); box_id++)
     {
         // grid points in box
 
@@ -1246,7 +1246,7 @@ integrateKxcFockForGGA(const std::vector<double*>& aoFockPointers,
 
     timer.stop("Preparation");
 
-    for (int box_id = 0; box_id < counts.size(); box_id++)
+    for (size_t box_id = 0; box_id < counts.size(); box_id++)
     {
         // grid points in box
 
@@ -1845,7 +1845,7 @@ integrateKxcLxcFockForGGA(const std::vector<double*>& aoFockPointers,
 
     timer.stop("Preparation");
 
-    for (int box_id = 0; box_id < counts.size(); box_id++)
+    for (size_t box_id = 0; box_id < counts.size(); box_id++)
     {
         // grid points in box
